Simple menu for KVM management
This is just a simple menu script for managing KVM.   Start , Shutdown and Connect to the console of the VM. 
It list all VMs.  
For the connect to console you need to edit the boot configuration to all the connection.  
edit /etc/default/grub add or modify line with GRUB_CMDLINE_LINUX_DEFAULT
GRUB_CMDLINE_LINUX_DEFAULT='console=ttySO xterm"
I think that is all.   you should be able to user the command "virsh console <vm_name>"  and you are connected to the VM console. 

Need to add a backup menu command and a way to end the script without using ctrl-c